+++ projects/angular-sdk/src/i18n/i18n.interface.ts
@@ -256,6 +256,67 @@ export interface I18n {
                 ZM: string;
                 ZW: string;
             },
+            STATE: {
+                AL: string;
+                AK: string;
+                AS: string;
+                AZ: string;
+                AR: string;
+                CA: string;
+                CO: string;
+                CT: string;
+                DE: string;
+                DC: string;
+                FM: string;
+                FL: string;
+                GA: string;
+                HI: string;
+                ID: string;
+                IL: string;
+                IN: string;
+                IA: string;
+                KS: string;
+                KY: string;
+                LA: string;
+                ME: string;
+                MH: string;
+                MD: string;
+                MA: string;
+                MI: string;
+                MN: string;
+                MS: string;
+                MO: string;
+                MT: string;
+                NE: string;
+                NV: string;
+                NH: string;
+                NJ: string;
+                NM: string;
+                NY: string;
+                NC: string;
+                ND: string;
+                MP: string;
+                OH: string;
+                OK: string;
+                OR: string;
+                PA: string;
+                RI: string;
+                SC: string;
+                SD: string;
+                TN: string;
+                TX: string;
+                UT: string;
+                VT: string;
+                VI: string;
+                VA: string;
+                WA: string;
+                WV: string;
+                WI: string;
+                WY: string;
+                AA: string;
+                AP: string;
+                AE: string;
+            },
             BUTTON: {
                 CLOSE: string;
             },