+++ projects/angular-sdk/CHANGELOG.md
@@ -1,7 +1,9 @@
 <a name="6.0.7"></a>
 
 # 6.0.7 (2024-XX-XX)
-- Fix logging issue.
+
+- Fixed logging issue.
+- Fixed an issue with US states in the refund form
 
 <a name="6.0.6"></a>